% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkWordpressXMLSitemaps.R
\name{checkWordpressXMLSitemaps}
\alias{checkWordpressXMLSitemaps}
\title{checkWordpressXMLSitemaps}
\usage{
checkWordpressXMLSitemaps(domain)
}
\arguments{
\item{domain}{domain name string}
}
\value{
data frame
}
\description{
check classic wordpress urls http code and return it inside the original data frame
}
