% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXMLSitemap.R
\name{getXMLSitemap}
\alias{getXMLSitemap}
\title{getXMLSitemap}
\usage{
getXMLSitemap(urltocheck)
}
\arguments{
\item{urltocheck}{direct xml sitemap url or hostname string of the website you want to find xml sitemap from}
}
\value{
dataframe
}
\description{
getXMLSitemap
}
