% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXMLSitemapFromRobotsTxt.R
\name{getXMLSitemapFromRobotsTxt}
\alias{getXMLSitemapFromRobotsTxt}
\title{getXMLSitemapFromRobotsTxt
ceci est un description tres partiel}
\usage{
getXMLSitemapFromRobotsTxt(urltocheck)
}
\arguments{
\item{urltocheck}{hostname string of the website you want to find xml sitemap from robots}
}
\value{
string
}
\description{
getXMLSitemapFromRobotsTxt
ceci est un description tres partiel
}
