% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXMLSitemapHTTP.R
\name{getXMLSitemapHTTP}
\alias{getXMLSitemapHTTP}
\title{getXMLSitemapHTTP}
\usage{
getXMLSitemapHTTP(sitemap, user_agent)
}
\arguments{
\item{sitemap}{dataframe}

\item{user_agent}{User agent string to http request}
}
\value{
dataframe
}
\description{
Check if xml sitemap urls send a 200 http code
}
