% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsitemapCheckHTTP.R
\name{xsitemapCheckHTTP}
\alias{xsitemapCheckHTTP}
\title{xsitemapCheckHTTP}
\usage{
xsitemapCheckHTTP(sitemap)
}
\arguments{
\item{sitemap}{dataframe}
}
\value{
dataframe
}
\description{
Check if xml sitemap urls send a 200 http code
}
