% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsitemapCheckWordpress.R
\name{xsitemapCheckWordpress}
\alias{xsitemapCheckWordpress}
\title{xsitemapCheckWordpress}
\usage{
xsitemapCheckWordpress(domain)
}
\arguments{
\item{domain}{domain name string}
}
\value{
data frame
}
\description{
check classic wordpress urls http code and return it inside the original data frame
}
