% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsitemapGetFromRobotsTxt.R
\name{xsitemapGetFromRobotsTxt}
\alias{xsitemapGetFromRobotsTxt}
\title{xsitemapGetFromRobotsTxt
ceci est un description tres partiel}
\usage{
xsitemapGetFromRobotsTxt(urltocheck)
}
\arguments{
\item{urltocheck}{hostname string of the website you want to find xml sitemap from robots}
}
\value{
string
}
\description{
xsitemapGetFromRobotsTxt
ceci est un description tres partiel
}
