% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xsitemapGuess.R
\name{xsitemapGuess}
\alias{xsitemapGuess}
\title{Title xsitemapGuess}
\usage{
xsitemapGuess(urltocheck)
}
\arguments{
\item{urltocheck}{hostname string of the website you want to guess xml sitemap url}
}
\value{
dataframe
}
\description{
surement la description la plus courte du monde
????
}
\details{
??
}
