% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXMLSitemap.R
\name{getXMLSitemap}
\alias{getXMLSitemap}
\title{getXMLSitemap}
\usage{
getXMLSitemap(urltocheck, user_agent)
}
\arguments{
\item{urltocheck}{direct xml sitemap url or hostname string of the website you want to find xml sitemap from}

\item{user_agent}{User agent string to http request}
}
\value{
dataframe
}
\description{
getXMLSitemap
}
