% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXMLSitemapHTTP.R
\name{getXMLSitemapHTTP}
\alias{getXMLSitemapHTTP}
\title{getXMLSitemapHTTP}
\usage{
getXMLSitemapHTTP(sitemap)
}
\arguments{
\item{sitemap}{dataframe}
}
\value{
dataframe
}
\description{
Check if xml sitemap urls send a 200 http code
}
