% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessXMLSitemap.R
\name{guessXMLSitemap}
\alias{guessXMLSitemap}
\title{Title guessXMLSitemap}
\usage{
guessXMLSitemap(urltocheck)
}
\arguments{
\item{urltocheck}{hostname string of the website you want to guess xml sitemap url}
}
\value{
dataframe
}
\description{
surement la description la plus courte du monde
????
}
\details{
??
}
